+++ app/src/main/java/woowacourse/shopping/ui/productList/event/ProductListNavigationEvent.kt
@@ -1,4 +1,4 @@
-package woowacourse.shopping.ui.productList
+package woowacourse.shopping.ui.productList.event
 
 sealed class ProductListNavigationEvent {
     data class ProductDetail(val productId: Long) : ProductListNavigationEvent()
