+++ app/src/main/java/woowacourse/shopping/ui/productList/viewholder/ProductsItemViewHolder.kt
@@ -1,4 +1,4 @@
-package woowacourse.shopping.ui.productList
+package woowacourse.shopping.ui.productList.viewholder
 
 import androidx.recyclerview.widget.RecyclerView
 import woowacourse.shopping.databinding.HolderProductBinding
